// lib/gen/api/discourseAPI/schemas/getTopic200PostStreamPostsItem.ts
export type GetTopic200PostStreamPostsItem = {
  id: number;
  name: string;
  username: string;
  avatar_template: string;
  created_at: string;
  cooked: string;
  post_number: number;
  post_type: number;
  updated_at: string;
  reply_count: number;
  reply_to_post_number: GetTopic200PostStreamPostsItemReplyToPostNumber;
  quote_count: number;
  incoming_link_count: number;
  reads: number;
  readers_count: number;
  score: number;
  yours: boolean;
  topic_id: number;
  topic_slug: string;
  display_username: string;
  primary_group_name: GetTopic200PostStreamPostsItemPrimaryGroupName;
  flair_name: GetTopic200PostStreamPostsItemFlairName;
  flair_url: GetTopic200PostStreamPostsItemFlairUrl;
  flair_bg_color: GetTopic200PostStreamPostsItemFlairBgColor;
  flair_color: GetTopic200PostStreamPostsItemFlairColor;
  version: number;
  can_edit: boolean;
  can_delete: boolean;
  can_recover: boolean;
  can_see_hidden_post?: boolean;
  can_wiki: boolean;
  link_counts: GetTopic200PostStreamPostsItemLinkCountsItem[];
  read: boolean;
  user_title: GetTopic200PostStreamPostsItemUserTitle;
  bookmarked: boolean;
  actions_summary: GetTopic200PostStreamPostsItemActionsSummaryItem[];
  moderator: boolean;
  admin: boolean;
  staff: boolean;
  user_id: number;
  hidden: boolean;
  trust_level: number;
  deleted_at: GetTopic200PostStreamPostsItemDeletedAt;
  user_deleted: boolean;
  edit_reason: GetTopic200PostStreamPostsItemEditReason;
  can_view_edit_history: boolean;
  wiki: boolean;
  reviewable_id: number;
  reviewable_score_count: number;
  reviewable_score_pending_count: number;
};